requirements.txt
numpy
torch
gymnasium
imageio==2.34.0
pygame
PyYAML
